+++ internal/modulecontext/from_proto.go
